+++ CHANGES.rst
@@ -12,11 +12,19 @@ Changes
 
 Changes:
 --------
-- No change.
+- Add `Job` status `HTML` response (resolves `#779 <https://github.com/crim-ca/weaver/issues/779>`_).
+- Add the ``process`` property to `Job` status response when requesting ``profile=openEO``,
+  with a direct reference to the underlying `CWL` `Application Package` of the main `Process` ran by the `Job`.
 
 Fixes:
 ------
-- No change.
+- Fix `W3C PROV` endpoints not returning contents in appropriate type when using ``f`` or ``format`` query parameter.
+  On top of the supported explicit ``Accept`` header, the endpoints will now also allow either explicit ``Content-Type``
+  passed by ``f`` / ``format`` query parameter, their shorthand representations (e.g.: ``json`` for ``application/json``
+  and their more verbose ``PROV``-specific representation (e.g.: ``f=prov-n``), all case-insensitive.
+- Fix ``/prov`` endpoint not correctly allowing the `YAML` equivalent representation of ``PROV-JSON`` contents.
+- Fix reported ``$schema`` to point at the `openEO` *Batch Job* `OenAPI` definition when requesting ``profile=openEO``.
+- Fix `Job` statistics not reported by the API in case of execution failure, although they might be partially available.
 
 .. _changes_6.4.1: